# Optimize BITCOUNT using ARM NEON SIMD