Replaced scalar loop with ARM NEON intrinsics for vectorized processing on ARM.

Results:

**Throughout** 
| Payload Size | Scalar Throughput (k req/s) | SIMD Throughput (k req/s) | Improvement (%) |
|--------------|-----------------------------|----------------------------|-----------------|
| 16B          | 249.69                      | 249.69                     | 0.00%           |
| 256B         | 249.63                      | 249.69                     | +0.02%          |
| 4KB          | 199.72                      | 249.63                     | +25.00%         |
| 64KB         | 44.33                       | 166.42                     | +275.43%        |
| 1MB          | 3.30                        | 26.59                      | +705.74%        |
| 10MB         | 0.33                        | 3.32                       | +900.04%        |


**Average Latency**


| Payload Size | Scalar Avg Latency (ms) | SIMD Avg Latency (ms) |
|--------------|--------------------------|-------------------------|
| 16B          | 0.374                    | 0.375                   |
| 256B         | 0.381                    | 0.376                   |
| 4KB          | 0.489                    | 0.389                   |
| 64KB         | 2.241                    | 0.575                   |
| 1MB          | 30.169                   | 3.649                   |
| 10MB         | 287.228                  | 29.220                  |


**P99 Latency**
| Payload Size | Scalar p99 Latency (ms) | SIMD p99 Latency (ms) |
|--------------|--------------------------|-------------------------|
| 16B          | 0.511                    | 0.511                   |
| 256B         | 0.519                    | 0.511                   |
| 4KB          | 0.639                    | 0.535                   |
| 64KB         | 2.439                    | 0.727                   |
| 1MB          | 32.303                   | 3.959                   |
| 10MB         | 314.623                  | 31.615                  |



Tested on AWS Graviton2.
To isolate CPU-bound improvements, the same key was used, reducing the likelihood of memory stalls for small payloads.

Fixes: #1864