@zuiderkwast 2025-05-11
Maybe I approved too early. :D Just one minor idea: In #1741 we added a unit test to check that the AVX and the scalar variant gave the same result. We could do the same for NEON. It's very easy to add it.