@xbasel 2025-05-11
Raw results:

cmd:
```
valkey-benchmark -h <ip> -n 10000 -c 100 --threads 50   bitcount key<size>
```

```

Before:

16bytes
Summary:
  throughput summary: 249687.89 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.374     0.112     0.383     0.495     0.511     0.791

256bytes:
Summary:
  throughput summary: 249625.56 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.381     0.128     0.391     0.495     0.519     0.711

4k:
Summary:
  throughput summary: 199720.39 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.489     0.136     0.495     0.615     0.639     0.847


64k:
Summary:
  throughput summary: 44330.17 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        2.241     0.216     2.271     2.399     2.439     6.655


1m:
Summary:
  throughput summary: 3296.41 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       30.169     0.792    30.975    31.903    32.303    35.871

10m:
Summary:
  throughput summary: 332.34 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
      287.228     3.232   305.663   314.623   314.623   409.343


After SIM optimization:
16bytes:
Summary:
  throughput summary: 249687.89 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.375     0.096     0.383     0.495     0.511     0.735

256bytes:
Summary:
  throughput summary: 249687.89 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.376     0.112     0.383     0.495     0.511     0.663

4k:
Summary:
  throughput summary: 249625.56 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.389     0.128     0.391     0.503     0.535     0.791

64k:
Summary:
  throughput summary: 166417.05 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        0.575     0.136     0.575     0.703     0.727     1.007


1M:
Summary:
  throughput summary: 26588.67 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        3.649     0.240     3.743     3.903     3.959     4.119


10M:
Summary:
  throughput summary: 3322.26 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       29.220     0.736    30.143    31.071    31.615    34.623

```