@xbasel 2025-05-11
> LGTM.
> 
> What size did you use in the benchmark? I guess the it depends on size just like #1741, right?
I've updated the description with up to date tests.